Feature - Playable Character:
Character movement now follows terrain constraints :
- Character can't go out of bounds
- Character can't jump off of cliffs
- Character can't climb cliffs
- Character can take stairs/slopes up or down
- Character sprite is displayed with an offset on stairs/slopes (half of a tile height)

Validation of the movement is moved from Character to Game, as Character isn't in charge of the map, and thus doesn't have access to the tiles' info.
This all may move to a map class, out of the game class, later.
As such, get_orientation_tile_sprite_id() has also been moved to Game.

support.get_sprite_rect() has been changes to not include Character-specific changes on a boolean, these changes are now properly included in character.get_rectangle(), that also takes into account if the player is on stairs/slopes or not.

Caveats :
- Character sprite orientation doesn't change with direction change.